moved halo implementation to dash/experimental
