+++ dash/test/GlobStencilIterTest.h
@@ -3,7 +3,8 @@
 
 #include <gtest/gtest.h>
 #include <libdash.h>
-
+#include <dash/experimental/Halo.h>
+#include <dash/experimental/iterator/GlobStencilIter.h>
 /**
  * Test fixture for class dash::GlobStencilIter.
  */
